Move procrustes.py from pose3d package to root
